+++ packages/dialect-react-sdk/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [UNRELEASED]
 
 - feature: add pluggable identity providers
+- feature: add the total amount of unread messages globally and per-thread
 
 ## [1.0.0-beta.26] - 2022-08-17